Avoid circular dependency on Color so that AMD compiles are valid
