+++ src/colorFn.js
@@ -1,0 +1,10 @@
+export function Color() {};
+
+Color.prototype = {
+  displayable: function() {
+    return this.rgb().displayable();
+  },
+  toString: function() {
+    return this.rgb() + "";
+  }
+};
